Sketch new protocol handler in `transport` module
